c03_reclear

+++ c03/ex01/ft_strncmp.c
@@ -12,10 +12,10 @@
 
 int	ft_strncmp(char *s1, char *s2, unsigned int n)
 {
-	int i;
+	unsigned int i;
 
 	i = 0;
-	if (n == 0)
+	if (n == (unsigned int)0)
 		return (0);
 	while (i < n && s1[i] == s2[i] && s1[i] != '\0' && s2[i] != '\0')
 		i++;
